Renamed solution.
