+++ src/Fluxera.Guards.UnitTests/AgainstPredicateTests.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluxera.Guard.UnitTests
+﻿namespace Fluxera.Guards.UnitTests
 {
 	using System;
 	using System.Collections.Generic;
