+++ src/Fluxera.Guards.UnitTests/TestEnum.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluxera.Guard.UnitTests
+﻿namespace Fluxera.Guards.UnitTests
 {
 	public enum TestEnum : int
 	{
