+++ src/Fluxera.Guards/Against.Null.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluxera.Guard
+﻿namespace Fluxera.Guards
 {
 	using System;
 	using System.Collections.Generic;
