+++ src/Fluxera.Guards/Guard.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluxera.Guard
+﻿namespace Fluxera.Guards
 {
 	using JetBrains.Annotations;
