+++ src/Fluxera.Guards/IGuard.cs
@@ -1,4 +1,4 @@
-﻿namespace Fluxera.Guard
+﻿namespace Fluxera.Guards
 {
 	using JetBrains.Annotations;
