+++ src/Fluxera.Guards/Properties/AssemblyInfo.cs
